refactor(gamelift-build): convert to libnuke resource format
